moves common stuff to a common folder
