+++ www/main/addons/bani-controller/components/ConnectionSwitch.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { remote } from 'electron';
-import { Switch } from '../../../sttm-ui';
+import { Switch } from '../../../common/sttm-ui';
 
 const analytics = remote.getGlobal('analytics');
