+++ www/main/common/constants/global-init.js
