+++ www/main/common/store/GlobalState.js
