+++ www/main/common/sttm-ui/switch/index.js
